+++ docs/toDO.txt
@@ -14,6 +14,9 @@ TODO SERVER/ CLIENT/ CHANNEL /ELSE:
 
 
 17.07
+BRANCH: "refactor/separate-invite-commands"l 
+	separated invite from commandsChannel
+
 BRANCH: refactor/separate-mode-from-commands-channel 
 	separated mode from commandsChannel
